[TASK] Build with younger inkscape
Inkscape 0.92.5 (2060ec1f9f, 2020-04-08)

It seems recent inkscape dropped support for
--export-filename in favor of more specific
--export-<type>:

Invalid option --export-filename=/.../out/cover/back.pdf

The patch adapts build.go to deal with this.

This is just a "works on my machine", though:
It would probably be better to adapt the build
chain to properly deal with different inkscape
versions.